% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_est_functions.R
\name{KDE_est}
\alias{KDE_est}
\title{Generate the Kernel Density Estimation results}
\usage{
KDE_est(x, grid = NULL, h, ker = "normal")
}
\arguments{
\item{x}{numeric vector, the samples;}

\item{grid}{numeric vector, the grid points;}

\item{h}{positive numeric, the bandwidth;}

\item{ker}{character, name of the kernel function.}
}
\value{
list of esimated densities and corresponding grid points.
}
\description{
Generate the Kernel Density Estimation results
}
\examples{
KDE_est(x=rnorm(50),grid=seq(-2,2, length.out = 100),h=NULL,ker="normal")
}
