% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_est_functions.R
\name{kde_est_big}
\alias{kde_est_big}
\title{Kernel density estimation simulations over given sample size, kernel, and bandwidth}
\usage{
kde_est_big(x, n, ker, h, grid)
}
\arguments{
\item{x}{numeric vector, the large population;}

\item{n}{positive numeric, the sample size;}

\item{ker}{character, name of the kernel function;}

\item{h}{positive numeric, the bandwidth;}

\item{grid}{numeric vector, the grid points.}
}
\value{
vector of the estimated densities.
}
\description{
Kernel density estimation simulations over given sample size, kernel, and bandwidth
}
\examples{
kde_est_big(x=rnorm(500), n=100,ker="normal", h=0.5, grid=seq(-2,2, length.out = 100))
}
