% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_est_functions.R
\name{sim_big_made}
\alias{sim_big_made}
\title{Kernel density estimation simulations over differnt combinations of sample sizes, kernels, and bandwidths}
\usage{
sim_big_made(x, ns, kers, hs, grid, true_f)
}
\arguments{
\item{x}{numeric vector, the large population;}

\item{ns}{numeric vector, different sample sizes;}

\item{kers}{character vector, names of different kernel functions;}

\item{hs}{positive numeric data.frame, different bandwidths;}

\item{grid}{numeric vector, the grid points;}

\item{true_f}{}
}
\value{
list of MADEs under different combination of parameters.
}
\description{
Kernel density estimation simulations over differnt combinations of sample sizes, kernels, and bandwidths
}
\examples{
# parameter setting
x <- rnorm(500)
ns <- c(100, 200)
kers <- c("normal","biweight")
hs <- data.frame(h1=0.1,h2=0.5)
grid <- seq(-2,2, length.out = 100)
true_f <- dnorm(grid)

sim_big_made(x, ns, kers, hs, grid, true_f)
}
